refactor: move conversation label helpers
